feat(code-example-evaluator): add loading state to evaluations display
Update the evaluations section to show a loading indicator while evaluations
are being fetched. Pass isLoading from the page component to the evaluations
section and conditionally render a loading message instead of evaluation cards.

Remove unused code and imports in the route related to filtering and loading
evaluations, simplifying evaluation data handling and improving UI responsiveness.
